requirements.txt
aiogram==2.13
python-dotenv==0.17.1
celery==5.1.1
click==7.1.2
redis==3.5.3
uniswap-python==0.5.1
requests==2.25.1
